Compartmentalized status component into individual file
